requirements.txt
asgiref==3.11.0
cffi==2.0.0
colorama==0.4.6
cryptography==46.0.3
Django==6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
pillow==12.0.0
pycparser==2.23
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
sqlparse==0.5.4
tzdata==2025.3
uritemplate==4.2.0
